% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzschleuder.R
\name{graph_from_netzschleuder}
\alias{graph_from_netzschleuder}
\title{Create a graph from the Netzschleuder data catalogue}
\usage{
graph_from_netzschleuder(name)
}
\arguments{
\item{name}{character. name of the network dataset. To get a network from a collection, use \verb{<collection_name>/<network_name>}.}
}
\value{
a new graph object.
}
\description{
Netzschleuder (\url{https://networks.skewed.de/}) is a large online repository for
network datasets with the aim of aiding scientific research.
}
