% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraphdata-package.R
\docType{package}
\name{igraphdata-package}
\alias{igraphdata}
\alias{igraphdata-package}
\title{igraphdata: A Collection of Network Data Sets for the 'igraph' Package}
\description{
A small collection of various network data sets, to use with the 'igraph' package: the Enron email network, various food webs, interactions in the immunoglobulin protein, the karate club network, Koenigsberg's bridges, visuotactile brain areas of the macaque monkey, UK faculty friendship network, domestic US flights network, etc. Also provides access to the API of \url{https://networks.skewed.de/}.
}
\section{How to use the data sets}{

After loading the \pkg{igraphdata} package, the various data sets are available by accessing them directly.

Get a list of data sets included in this package with \code{data(package = "igraphdata")}.
}

\seealso{
Useful links:
\itemize{
  \item \url{http://igraph.org}
  \item Report bugs at \url{https://github.com/igraph/igraphdata/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{kirill@cynkra.com} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})

Authors:
\itemize{
  \item Kirill Müller \email{kirill@cynkra.com} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})
  \item Gábor Csárdi (\href{https://orcid.org/0000-0001-7098-9676}{ORCID})
  \item David Schoch (\href{https://orcid.org/0000-0003-2952-4812}{ORCID})
}

Other contributors:
\itemize{
  \item Szabolcs Horvát (\href{https://orcid.org/0000-0002-3100-523X}{ORCID}) [contributor]
  \item Maëlle Salmon (\href{https://orcid.org/0000-0002-2815-0399}{ORCID}) [contributor]
  \item Chan Zuckerberg Initiative [funder]
}

}
\keyword{internal}
