% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzschleuder.R
\name{ns_metadata}
\alias{ns_metadata}
\alias{ns_df}
\alias{ns_graph}
\title{Download and Convert Graph Data from Netzschleuder}
\usage{
ns_metadata(name, collection = FALSE)

ns_df(name, token = NULL, size_limit = 1)

ns_graph(name, token = NULL, size_limit = 1)
}
\arguments{
\item{name}{Character. The name of the network dataset.
To get a network from a collection, use the format \verb{<collection_name>/<network_name>}.}

\item{collection}{Logical. If TRUE, get the metadata of a whole collection of networks.}

\item{token}{Character. Some networks have restricted access and require a token.}

\item{size_limit}{Numeric. Maximum allowed file size in GB.
Larger files will be prevented from being downloaded.
See \url{https://networks.skewed.de/restricted}.}
}
\value{
\describe{
\item{\code{ns_metadata()}}{A list containing metadata for the dataset.}
\item{\code{ns_df()}}{A named list with \code{nodes}, \code{edges}, \code{gprops}, and \code{meta}.}
\item{\code{ns_graph()}}{An \code{igraph} object.}
}
}
\description{
These functions provide tools to interact with the Netzschleuder network dataset archive.
Netzschleuder (\url{https://networks.skewed.de/}) is a large online repository for network datasets,
aimed at aiding scientific research.
\describe{
\item{\code{ns_metadata()}}{ retrieves metadata about a network or network collection.}
\item{\code{ns_df()}}{downloads the graph data as data frames (nodes, edges, and graph properties).}
\item{\code{ns_graph()}}{creates an \code{igraph} object directly from Netzschleuder.}
}
}
\examples{
\dontrun{
# Get metadata
ns_metadata("copenhagen/calls")

# Download network as data frames
graph_data <- ns_df("copenhagen/calls")

# Create an igraph object
g <- ns_graph("copenhagen/calls")
}

}
\seealso{
\url{https://networks.skewed.de/}
}
