\docType{package}
\name{igraphdata-package}
\alias{igraphdata-package}
\alias{igraphdata}
\title{The igraphdata package}
\description{The igraphdata package provides various data sets }
\section{How to use the data sets}{
  After loading the \code{igraphdata} package, the various data sets can
  be loaded using the regular \code{\link{data}} command.

  Type in \preformatted{  data(package="igraphdata")}
  to get a list of data sets included in this package.
}
\author{Gabor Csardi \email{csardi.gabor@gmail.com}}
\references{\code{\link{data}}}
%\examples{}
