% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzschleuder.R
\name{graph_from_netzschleuder}
\alias{graph_from_netzschleuder}
\title{Create a graph from the Netzschleuder data catalogue}
\usage{
graph_from_netzschleuder(
  name,
  ...,
  network = NULL,
  directed = FALSE,
  bipartite = FALSE
)
}
\arguments{
\item{name}{character. name of the network dataset.}

\item{network}{character. If the dataset contains several networks this is the network name.}

\item{directed}{logical. Whether a directed graph is constructed.}

\item{bipartite}{logical. Whether a bipartite graph is constructed.}
}
\value{
a new graph object.
}
\description{
Netzschleuder (\url{https://networks.skewed.de/}) is a large online repository for
network datasets with the aim of aiding scientific research.
}
