% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{igraphdata-package}
\alias{igraphdata-package}
\alias{igraphdata}
\title{The igraphdata package}
\description{
The igraphdata package provides various data sets
}
\section{How to use the data sets}{

After loading the \code{igraphdata} package, the various data sets can
be loaded using the regular \code{\link{data}} command.

Type in \preformatted{  data(package="igraphdata")}
to get a list of data sets included in this package.
}

\references{
\code{\link{data}}
}
\author{
Gabor Csardi \email{csardi.gabor@gmail.com}
}
