% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraphdata-package.R
\docType{package}
\name{igraphdata-package}
\alias{igraphdata}
\alias{igraphdata-package}
\title{igraphdata: A Collection of Network Data Sets for the 'igraph' Package}
\description{
A small collection of various network data sets, to use with the 'igraph' package: the Enron email network, various food webs, interactions in the immunoglobulin protein, the karate club network, Koenigsberg's bridges, visuotactile brain areas of the macaque monkey, UK faculty friendship network, domestic US flights network, etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://igraph.org/}
  \item Report bugs at \url{https://github.com/igraph/igraphdata/issues}
}

}
\keyword{internal}
