% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuno.R
\docType{data}
\name{immuno}
\alias{immuno}
\title{Immunoglobulin interaction network}
\format{
An undirected \code{igraph} graph object.

Graph attributes: \sQuote{name}, \sQuote{Citation}, \sQuote{Author}.
}
\source{
See reference below.
}
\usage{
immuno
}
\description{
The undirected and connected network of interactions
in the immunoglobulin protein. It is made up of 1316 vertices
representing amino-acids and an edge is drawn between two
amino-acids if the shortest distance between their C_alpha atoms
is smaller than the threshold value \eqn{\theta=8}{theta=8} Angstrom.
}
\references{
D. Gfeller, Simplifying complex networks: from a clustering to a
coarse graining strategy, \emph{PhD Thesis EPFL}, no 3888, 2007.
\url{https://infoscience.epfl.ch/entities/publication/35db87cd-756f-42c5-b0a5-60f4de0eae69}
}
\keyword{datasets}
