\name{kite}
\docType{data}
\alias{kite}
\title{Krackhardt's kite}
\description{
  Krackhardt's kite is a fictionary social network with ten actors.
  It is the smallest graph for which the most central actor is different
  according to the three classic centality measures: degree, closeness
  and betweenness.
}
\usage{
kite
}
\format{
  An undirected igraph graph with graph attributes \code{name},
  \code{layout}, \code{Citation}, \code{Author}, \code{URL}, and vertex
  attributes \code{label}, \code{name} and \code{Firstname}.
}
\source{
  \url{http://www.orgnet.com/sna.html}
}
\references{
  Assessing the Political Landscape: Structure, Cognition, and Power in
  Organizations. David Krackhardt. \emph{Admin. Sci. Quart.} 35, 342-369,
  1990.
}
\keyword{datasets}
