% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kite.R
\docType{data}
\name{kite}
\alias{kite}
\title{Krackhardt's kite}
\format{
An undirected igraph graph with graph attributes \code{name}, \code{layout}, \code{Citation}, \code{Author}, \code{URL},
and vertex attributes \code{label}, \code{name} and \code{Firstname}.
}
\source{
\url{http://www.orgnet.com/sna.html}
}
\usage{
kite
}
\description{
Krackhardt's kite is a fictional social network with ten actors.
 It is a small (though not the smallest possible) graph for which the most central actors are different
according to the three classic centrality measures: degree, closeness and betweenness.
}
\references{
Assessing the Political Landscape: Structure, Cognition, and Power in Organizations.
 David Krackhardt. \emph{Admin. Sci. Quart.} 35, 342-369, 1990. \doi{10.2307/2393394}
}
\keyword{datasets}
