% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{lesmis}
\alias{lesmis}
\alias{lesmis_gml}
\alias{lesmis_graphml}
\alias{lesmis_pajek}
\title{Example files}
\usage{
lesmis_gml()

lesmis_graphml()

lesmis_pajek()
}
\value{
A string indicating an absolute path to a file.
}
\description{
Functions that return paths to example files of the "Les Miserables" example network,
in the GML, GraphML or Pajek format.
}
