% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macaque.R
\docType{data}
\name{macaque}
\alias{macaque}
\title{Visuotactile brain areas and connections}
\format{
A directed \code{igraph} graph object with vertex attributes
\sQuote{name} and \sQuote{shape}.

This dataset is licensed under a Creative Commons
Attribution-Share Alike 2.0 UK: England & Wales License,
see \url{http://creativecommons.org/licenses/by-sa/2.0/uk/} for details.
Please cite the reference below if you use this dataset.
}
\source{
See reference below.
}
\usage{
macaque
}
\description{
Graph model of the visuotactile brain areas and connections of the
macaque monkey. The model consists of 45 areas and 463 directed
connections.
}
\references{
Negyessy L., Nepusz T., Kocsis L., Bazso F.: Prediction of
the main cortical areas and connections involved in the tactile
function of the visual cortex by network analysis. \emph{European
Journal of Neuroscience}, 23(7): 1919-1930, 2006.
}
\keyword{datasets}
