% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzschleuder.R
\name{read_from_netzschleuder}
\alias{read_from_netzschleuder}
\title{Download a graph from the Netzschleuder data catalogue
Netzschleuder (\url{https://networks.skewed.de/}) is a large online repository for
network datasets with the aim of aiding scientific research.}
\usage{
read_from_netzschleuder(name, token = NULL)
}
\arguments{
\item{name}{character. name of the network dataset. To get a network from a collection, use \verb{<collection_name>/<network_name>}.}

\item{token}{character. Some networks have restricted access and need a toke. See \url{https://networks.skewed.de/restricted}}
}
\value{
a named list containing an edge list and node attribute data frame and some metadata
}
\description{
Download a graph from the Netzschleuder data catalogue
Netzschleuder (\url{https://networks.skewed.de/}) is a large online repository for
network datasets with the aim of aiding scientific research.
}
